refactor(main ui component): moved all data formatters to sub folder , for clarity
