+++ src/components/main/dataFormating/formatDataForLocalStorage.js
